requirements.txt
# CARLA-Alpamayo Integration Requirements

# Core dependencies
numpy>=1.21.0
Pillow>=9.0.0

# PyTorch (install with CUDA support)
# torch>=2.0.0  # Install separately with: pip install torch --index-url https://download.pytorch.org/whl/cu118

# Transformers for Alpamayo model
transformers>=4.36.0
accelerate>=0.25.0
huggingface_hub>=0.20.0

# CARLA Python API
# carla  # Install from CARLA package: pip install carla

# Optional: Flash Attention 2 (for faster inference)
# flash-attn>=2.0.0  # Requires CUDA toolkit

# Visualization and analysis
matplotlib>=3.5.0
opencv-python>=4.5.0
tqdm>=4.62.0

# Recording and data handling
pandas>=1.3.0
